+++ test/router/index.js
@@ -1081,11 +1081,22 @@ describe('gracenode.router', function () {
 		gn.router.static('/static', [ '../../../test/router/static/' ]);
 	});
 
-	it('can handle static file route', function (done) {
+	it('can handle static file route /static/test/router/static/spinner.gif', function (done) {
 		var opt = {
 			gzip: false
 		};
 		request.GET(http + '/static/test/router/static/spinner.gif', {}, opt, function (error, res, st) {
+			assert.equal(error, null);
+			assert.equal(st, 200);
+			done();
+		});
+	});
+
+	it('can handle static file route /static/test/router/static/more/spinner.gif', function (done) {
+		var opt = {
+			gzip: false
+		};
+		request.GET(http + '/static/test/router/static/more/spinner.gif', {}, opt, function (error, res, st) {
 			assert.equal(error, null);
 			assert.equal(st, 200);
 			done();