+++ src/router/hooks.js
@@ -22,7 +22,9 @@ exports.hook = function (path, func) {
 		} else {
 			hooks[path].push(func);
 		}
-		logger.verbose('HTTP request hook registed:', path, 'hooks #', hooks[path].length);
+		if (logger) {
+			logger.verbose('HTTP request hook registed:', path, 'hooks #', hooks[path].length);
+		}
 		return;
 	}
 	var headingSlash = path[0] === '/' ? '' : '/';
@@ -38,7 +40,9 @@ exports.hook = function (path, func) {
 	} else {
 		hooks[hookPath].push(func);
 	}
-	logger.verbose('HTTP request hook registed:', hookPath, 'hooks #', hooks[hookPath].length);
+	if (logger) {
+		logger.verbose('HTTP request hook registed:', hookPath, 'hooks #', hooks[hookPath].length);
+	}
 };
 
 exports.updateHooks = function (fastRoutes, routes, allroutes) {